userauth-ms/requirements.txt
Flask==2.3.3
python-dotenv==1.1.0
requests==2.31.0
PyJWT==2.8.0
werkzeug==2.3.7
pytest==7.4.0
Flask-Cors
passlib[bcrypt]